+++ navigation/lib-navigation-rx/src/main/java/ru/surfstudio/android/navigation/rx/extension/TabFragmentNavigatorExtensions.kt
@@ -7,6 +7,11 @@ import ru.surfstudio.android.navigation.navigator.fragment.tab.TabFragmentNaviga
 import ru.surfstudio.android.navigation.rx.ActiveTabReopenedObservable
 import ru.surfstudio.android.navigation.rx.FragmentBackStackChangedObservable
 
-fun TabFragmentNavigatorInterface.activeTabReopens(): Observable<String> {
+/**
+ * Get all events at which currently active and empty tab is reopened.
+ *
+ * @return [Observable] with current tab's tag.
+ */
+fun TabFragmentNavigatorInterface.observeActiveTabReopened(): Observable<String> {
     return ActiveTabReopenedObservable(this)
 }