+++ navigation/lib-navigation/src/main/java/ru/surfstudio/android/navigation/command/fragment/RemoveAll.kt
@@ -5,7 +5,6 @@ import ru.surfstudio.android.navigation.animation.DefaultAnimations
 import ru.surfstudio.android.navigation.command.fragment.base.FragmentNavigationCommand
 import ru.surfstudio.android.navigation.route.fragment.FragmentRoute
 import ru.surfstudio.android.navigation.route.stub.StubFragmentRoute
-import ru.surfstudio.android.navigation.route.stub.StubTabFragmentRoute
 
 
 /**
@@ -13,13 +12,16 @@ import ru.surfstudio.android.navigation.route.stub.StubTabFragmentRoute
  *
  * @param animations animations [Animations] used to specify navigation transition appearance
  * @param sourceTag tag of a source screen, which will execute navigation command.
+ * @param shouldRemoveLast flag which determines whether we should wipe all fragments in stack, or
+ * leave last one alive.
  * @param isTab are we removing fragment from tab navigator, or from default navigator
  */
 data class RemoveAll(
         override val animations: Animations = DefaultAnimations.fragment,
         override var sourceTag: String = "",
+        val shouldRemoveLast: Boolean = false,
         val isTab: Boolean = false
 ) : FragmentNavigationCommand {
 
-    override val route: FragmentRoute = if (isTab) StubTabFragmentRoute else StubFragmentRoute
+    override val route: FragmentRoute = StubFragmentRoute
 }