+++ navigation/lib-navigation/src/main/java/ru/surfstudio/android/navigation/command/fragment/RemoveLast.kt
@@ -1,12 +1,10 @@
 package ru.surfstudio.android.navigation.command.fragment
 
 import ru.surfstudio.android.navigation.animation.Animations
-import ru.surfstudio.android.navigation.animation.DefaultAnimations
 import ru.surfstudio.android.navigation.animation.resource.NoResourceAnimations
 import ru.surfstudio.android.navigation.command.fragment.base.FragmentNavigationCommand
 import ru.surfstudio.android.navigation.route.fragment.FragmentRoute
 import ru.surfstudio.android.navigation.route.stub.StubFragmentRoute
-import ru.surfstudio.android.navigation.route.stub.StubTabFragmentRoute
 
 /**
  * Remove last screen from current stack and switch to previous.
@@ -21,5 +19,5 @@ data class RemoveLast(
         val isTab: Boolean = false
 ) : FragmentNavigationCommand {
 
-    override val route: FragmentRoute = if (isTab) StubTabFragmentRoute else StubFragmentRoute
+    override val route: FragmentRoute = StubFragmentRoute
 }