+++ navigation/lib-navigation/src/main/java/ru/surfstudio/android/navigation/executor/screen/fragment/FragmentCommandExecutor.kt
@@ -6,7 +6,6 @@ import ru.surfstudio.android.navigation.command.fragment.base.FragmentNavigation
 import ru.surfstudio.android.navigation.provider.ActivityNavigationProvider
 import ru.surfstudio.android.navigation.executor.CommandExecutor
 import ru.surfstudio.android.navigation.navigator.fragment.FragmentNavigatorInterface
-import ru.surfstudio.android.navigation.route.tab.TabRoute
 
 /**
  * Command executor for [ActivityNavigationCommand]s.
@@ -20,12 +19,8 @@ open class FragmentCommandExecutor(
     override fun execute(command: FragmentNavigationCommand) {
         val activityHolder = activityNavigationProvider.provide()
         val fragmentNavigationHolder = activityHolder.fragmentNavigationProvider.provide(command.sourceTag)
-        val tabFragmentNavigator = fragmentNavigationHolder.tabFragmentNavigator
         val fragmentNavigator = fragmentNavigationHolder.fragmentNavigator
-        when (command.route) {
-            is TabRoute -> execute(command, tabFragmentNavigator)
-            else -> execute(command, fragmentNavigator)
-        }
+        execute(command, fragmentNavigator)
     }
 
     protected open fun execute(command: FragmentNavigationCommand, navigator: FragmentNavigatorInterface) {
@@ -36,7 +31,7 @@ open class FragmentCommandExecutor(
             is Remove -> navigator.remove(command.route, command.animations)
             is RemoveLast -> navigator.removeLast(command.animations)
             is RemoveUntil -> navigator.removeUntil(command.route, command.isInclusive)
-            is RemoveAll -> navigator.removeAll()
+            is RemoveAll -> navigator.removeAll(command.shouldRemoveLast)
         }
     }