+++ navigation/sample-standard/src/main/java/ru/surfstudio/android/navigation/sample_standard/di/ui/ActivityComponent.kt
@@ -1,0 +1,14 @@
+package ru.surfstudio.android.navigation.sample_standard.di.ui
+
+import dagger.Component
+import ru.surfstudio.android.dagger.scope.PerActivity
+import ru.surfstudio.android.navigation.sample_standard.di.AppComponent
+import ru.surfstudio.android.sample.dagger.ui.base.dagger.activity.DefaultActivityComponent
+import ru.surfstudio.android.sample.dagger.ui.base.dagger.activity.DefaultActivityModule
+
+@PerActivity
+@Component(
+        dependencies = [AppComponent::class],
+        modules = [DefaultActivityModule::class, ActivityNavigationModule::class]
+)
+interface ActivityComponent : DefaultActivityComponent, AppComponent