+++ navigation/RELEASE_NOTES.md
@@ -2,3 +2,4 @@
 
 ## 0.5.0-alpha
 ##### Navigation
+* ANDDEP-312 Added navigation module